requirements.txt
# Core Dependencies for Dairy Scheduler MVP
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.15.0
pyyaml>=6.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
numpy>=1.24.0
datetime
pathlib
typing
